% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindUD.R
\name{FindUD}
\alias{FindUD}
\title{FindUD}
\usage{
FindUD(
  InSeries = NULL,
  InCs,
  output,
  MissingValueCode = "-999.99",
  GUI = FALSE,
  plev = 0.95,
  Iadj = 10000,
  Mq = 10,
  Ny4a = 0,
  is_plot = FALSE
)
}
\arguments{
\item{InSeries}{file path of the input data}

\item{InCs}{A data.frame of detected turning points returned from \link{FindU},
with the columns of \code{c("kind", "Idc", "date", "probL", "probU", "plev", "PFx", "PFx95l", "PFx95h")}.}

\item{output}{prefix of the outputs}

\item{MissingValueCode}{string, missing value code}

\item{GUI}{boolean}

\item{plev}{plev is the nominal level of confidence; (1 - plev) is the
nominal level of significance, choose one of the following 6 plev values:
0.75, 0.80, 0.90, 0.95, 0.99, 0.9999}

\item{Iadj}{If Iadj = 0, the data series is adjusted to the longest segment;
otherwise the data series is adjusted to the chosen segment Iadj (if the
given integer Iadj > Ns+1, we re-set Iadj=Ns+1, which corresponds to
adjusting the series to the last segment). Set Iadj = 10000 if you want
to ensure that the series is adjusted to the last segment}

\item{Mq}{Mq (=0, 1, 2, ..., 20) is the number of points (categories) on which
the empirical cumulative distribution function are estimated.
If Mq=0, the actual Mq is determined by the length of the shortest segment
Mq=1 corresponds to mean adjustments (one adjustment for all data in
the same segment).}

\item{Ny4a}{}
}
\description{
FindUD
}
\examples{
f = system.file("extdata/Example1.dat", package = "RHtests")
r1 = FindU(f)

r2 = FindUD(InCs = r1$TP, output = "OUTPUT/example1")
print(r2)

plot_output(r2$fit)
}
