% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSmultiRed.R
\name{LSmultiRedCycle}
\alias{LSmultiRedCycle}
\title{LSmultiRedCycle}
\usage{
LSmultiRedCycle(Y0, Ti, Ips, Iseg.adj)
}
\value{
\itemize{
\item \code{resi}: residual
\item \code{meanhat}: linear trend
}
}
\description{
LSmultiRedCycle
}
