% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QMadjGaussian.wRef.R
\name{QMadjGaussian}
\alias{QMadjGaussian}
\alias{QMadjGaussian.wRef}
\title{QMadj}
\usage{
QMadjGaussian(P, Ips, Mq, Iseg.adj, Nadj)

QMadjGaussian.wRef(B, Dif, Ips, Mq, Iseg.adj, Nadj, Nt, Ny4a)
}
\value{
Example1_Ref1_DLY.dat_QMadjDLY_data.dat, which contains the original and the
adjusted daily data in its 4 th and 5 th column respectively.

(column 6, 7, 8 are the mean-adjusted daily series, the QM adjustments,
and the mean adjustments, respectively;

columns 4 + 7 = column 5;
columns 4 + 8 = column 6)
}
\description{
QMadj
}
