% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{mutate}
\alias{arrange}
\alias{write_fig}
\alias{data.table}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{Ipaper}{\code{\link[Ipaper]{write_fig}}}

  \item{data.table}{\code{\link[data.table]{data.table}}}

  \item{dplyr}{\code{\link[dplyr]{arrange}}, \code{\link[dplyr]{mutate}}}
}}

