% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHtests_process.R
\name{RHtests_process}
\alias{RHtests_process}
\title{process RHtests without reference series}
\usage{
RHtests_process(
  data,
  data.ref = NULL,
  metadata,
  prefix = NULL,
  outdir = "OUTPUT/example01",
  maxgap = 366,
  ...,
  debug = TRUE,
  is_plot = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{data}{A data.frame or data.table, with the columns of c('year',
'month', 'day', 'data')}

\item{data.ref}{A data.frame or data.table, with the columns of c('year',
'month', 'day', 'data'). The reference series of \code{data}, should have the same
length as \code{data}.}

\item{metadata}{A data.frame or data.table, with the columns of
\code{TurningPoint} date.}
}
\description{
process RHtests without reference series
}
\examples{
## example 01 ------------------------------------------------------------------
data <- dat_example01
metadata <- data.table(date = c("1966-11-01", "1976-07-01", "1980-03-01"))

r <- RHtests_process(data, NULL, metadata, outdir = "OUTPUT/example01")
plot_output(r$data)
## example 02 ------------------------------------------------------------------
Bseries <- system.file("extdata/Example2.dat", package = "RHtests")
Rseries <- system.file("extdata/Example2_Ref.dat", package = "RHtests")

metadata <- data.table(date = c("19740200", "19751100"))

r <- RHtests_process(Bseries, Rseries, metadata, outdir = "OUTPUT/example02")
plot_RHtests(r, "RHtests-ref.pdf")

# expect_equal(nrow(r$turningPoint), 11)
# expect_equal(as.character(r$turningPoint[kind == 0]$date),
#               metadata$date)
# expect_equal(nrow(r$data), 612)
}
