% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHtests_process.R
\name{RHtests_stepsize}
\alias{RHtests_stepsize}
\title{RHtests_stepsize}
\usage{
RHtests_stepsize(
  data = NULL,
  data.ref = NULL,
  TP2,
  has_ref = !is.null(data.ref),
  prefix = "./OUTPUT/example03",
  is_plot = FALSE,
  verbose = TRUE,
  ...,
  debug = TRUE
)
}
\description{
stepsize will not alter TP2, only delte non-significant record.
}
