% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StepSize.R
\name{StepSize}
\alias{StepSize}
\title{StepSize}
\usage{
StepSize(
  InSeries = NULL,
  InCs,
  output,
  MissingValueCode = "-999.99",
  GUI = FALSE,
  plev = 0.95,
  Iadj = 10000,
  Mq = 10,
  Ny4a = 0,
  is_plot = TRUE
)
}
\description{
StepSize
}
\examples{
## example 01 ------------------------------------------------------------------
data <- dat_example01
metadata <- data.table(date = c("1966-11-01", "1976-07-01", "1980-03-01"))
prefix <- "OUTPUT/example02/example01"
r <- RHtests_process(data, NULL, metadata, prefix)
plot_output(r$data)
}
