% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StepSize.wRef.R
\name{StepSize.wRef}
\alias{StepSize.wRef}
\title{StepSize.wRef}
\usage{
StepSize.wRef(
  Bseries = NULL,
  Rseries = NULL,
  InCs,
  output,
  MissingValueCode = "-999.99",
  plev = 0.95,
  Iadj = 10000,
  Mq = 10,
  GUI = F,
  Ny4a = 0,
  is_plot = TRUE
)
}
\description{
StepSize.wRef
}
\examples{
## example 02 ------------------------------------------------------------------
Bseries <- system.file("extdata/Example2.dat", package = "RHtests")
Rseries <- system.file("extdata/Example2_Ref.dat", package = "RHtests")

metadata <- data.table(date = c("19740200", "19751100"))
prefix <- "../../OUTPUT/example02/example02"

r <- process_RHtests(Bseries, Rseries, metadata, prefix)
plot_RHtests(r)

# expect_equal(nrow(r$turningPoint), 11)
# expect_equal(as.character(r$turningPoint[kind == 0]$date),
#               metadata$date)
# expect_equal(nrow(r$data), 612)
}
