% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TP_adjust.R
\name{TP_adjustByMeta}
\alias{TP_adjustByMeta}
\alias{TP_remove_least}
\title{adjust TP according to station meta info}
\usage{
TP_adjustByMeta(TP, metadata, maxgap = 366)

TP_remove_least(TP)
}
\arguments{
\item{TP}{A data.frame, with the columns of \code{c("kind", "diff", "")}}
}
\value{
An adjusted date
}
\description{
adjust TP according to station meta info
}
