% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindU.R
\name{FindU}
\alias{FindU}
\title{Penalized maximal F (PMF) test}
\usage{
FindU(
  InSeries = NULL,
  output = "./OUTPUT/example01",
  MissingValueCode = "-999.99",
  GUI = FALSE,
  plev = 0.95,
  Iadj = 10000,
  Mq = 10,
  Ny4a = 0,
  is_plot = FALSE,
  ...,
  debug = TRUE
)
}
\arguments{
\item{InSeries}{file path of the input data}

\item{output}{prefix of the outputs}

\item{MissingValueCode}{string, missing value code}

\item{GUI}{boolean}

\item{plev}{plev is the nominal level of confidence; (1 - plev) is the
nominal level of significance, choose one of the following 6 plev values:
0.75, 0.80, 0.90, 0.95, 0.99, 0.9999}

\item{Iadj}{If Iadj = 0, the data series is adjusted to the longest segment;
otherwise the data series is adjusted to the chosen segment Iadj (if the
given integer Iadj > Ns+1, we re-set Iadj=Ns+1, which corresponds to
adjusting the series to the last segment). Set Iadj = 10000 if you want
to ensure that the series is adjusted to the last segment}

\item{Mq}{Mq (=0, 1, 2, ..., 20) is the number of points (categories) on which
the empirical cumulative distribution function are estimated.
If Mq=0, the actual Mq is determined by the length of the shortest segment
Mq=1 corresponds to mean adjustments (one adjustment for all data in
the same segment).}

\item{Ny4a}{}
}
\value{
Example1_1Cs.txt, Example1_Ustat.txt, Example1_U.dat, Example1_U.pdf
\itemize{
\item Example1_1Cs.txt

Ns >= 0 is the number of changepoints identified; changepoint positions are
Ip(1), Ip(2),...,Ip(Ns)
\item OutFile_U.dat
}
\itemize{
\item \code{id}   :
\item \code{date} :
\item \code{base} : the original input series
\item \code{base_trend+meanshift}: \code{trend} + \verb{abrupt shift} (for base)
\item \code{base_mean_adj}: \code{base} - \verb{abrupt shift}
\item \code{base_anomaly}: \code{base} - \verb{seasonal cycle}
\item \code{fit_of_base_anomaly}: \code{trend} + \verb{abrupt shift} (for anomaly)
\item \code{annualC+meanShifts}: \verb{seasonal cycle} + \verb{abrupt shift} (for base)
\item \code{QM_adjusted}:
\item \code{fit_of_deseason_base}: linear trend of \code{anomal}. A multi-phase regression
model fit to the de-seasonalized base series without accounting for any
shifts (i.e. ignore all shifts identified)
}
}
\description{
PMF allows the time series being tested to have a linear trend throughout
the whole period of the data record (i.e., no shift in the trend component;
see Wang 2003), with the annual cycle, linear trend, and lag-1 autocorrelation
of the base series being estimated in tandem through iterative procedures,
while accounting for all the identified mean-shifts (Wang 2008a). No reference
series will be used in any of these functions; the base series is tested in
this case.
}
\examples{
f = system.file("extdata/Example1.dat", package = "RHtests")
r = FindU(f)
print(r)

plot_output(r$fit)

}
