% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSmultiRed.R
\name{LSmultiple}
\alias{LSmultiple}
\title{LSmultiple}
\usage{
LSmultiple(Y, T, Ips)
}
\arguments{
\item{Ips}{what?}
}
\value{
\itemize{
\item \code{sig}   : coefficients
\item \code{fitted}: fitted value
\item \code{resi}  : residual of fitted value
}
}
\description{
LSmultiple
}
