% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.R
\name{Read}
\alias{Read}
\title{read input of RHtestsV4}
\usage{
Read(InSeries, MissingValueCode = "-999.99")
}
\arguments{
\item{InSeries}{file path of the input data}

\item{MissingValueCode}{string, missing value code}
}
\value{
\itemize{
\item \code{lon}: name
\item \code{lat}: name
}
}
\description{
read input of RHtestsV4
}
\examples{
infile = system.file("extdata/Example1.dat", package = "RHtests")
Read(infile, "-999.99")
}
