% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.R
\name{Read}
\alias{Read}
\title{read input of RHtestsV4}
\usage{
Read(InSeries, MissingValueCode = "-999.99")
}
\arguments{
\item{InSeries}{file path of the input data}

\item{MissingValueCode}{string, missing value code}
}
\value{
If failed, NULL will be returned. If success, a data.table with the following
columns will be returned:
\itemize{
\item \code{IY0}: numeric, \code{year*1000 + month*100 + day}, e.g. '19980100'
\item \code{IY0flg}: 0 or 1, is y\link{t} has y\link{t+1} for autocorrelation?
\item \code{Imd}: numeric, \code{month*100 + day}, e.g. '100'
\item \code{Ti}: index or original Y
\item \code{Y0}: original Y after eliminating NA values
}

\code{NT}=1, 12, and 365 for annual, monthly, and daily series, respectively.
}
\description{
Remove NA values and flag y\link{t} as 0 when it has not corresponding y\link{t+1} values.
}
\examples{
infile = system.file("extdata/Example1.dat", package = "RHtests")
Read(infile, "-999.99")
}
